+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Changelog 
+
+- OpenCode: Install ripgrep for access to native skills.
+
 ## 0.2.55 (16 May 2026)
 
 - Workaround for `agent_client_protocol` v0.10.0 connection bug.